moved dna_construct into a different file
