+++ biocrnpyler/__init__.py
@@ -17,3 +17,4 @@ from .crnlab import *
 # Additional functions
 from .sbmlutil import *
 from .plotting import *
+from .dna_construct import *
